Refactor: NetworkManager 생성, 네트워크 통신 로직 분리
